feat(sdk): Support for the OAuth Discoverable Client ID Scheme
* Support for the OAuth Discoverable Client ID Scheme described in https://mattrglobal.github.io/draft-looker-oauth-client-id-scheme/draft-looker-oauth-client-id-scheme.html.
* Added documentation for how to use this scheme with Wallet-SDK.
* Updated documentation for the other methods of getting/providing a client ID.
* Eliminated some code duplication by doing some refactoring.
